A web page should one have one banner landmark
The role="banner" ARIA landmark should only appear once on an element that is unique to the HTML document even though the HMTL5 specification allows multiple instances of an equivalent header landmark value. It is appropriate in most web designs to have only one banner landmark whether it is specified using ARIA or HTML5.

https://dequeuniversity.com/rules/axe/3.3/landmark-no-duplicate-banner